rename buildbot.db.buildslave{,s} db.worker{,s}
as it should be.
