+++ master/buildbot/db/connector.py
@@ -20,7 +20,7 @@ from buildbot.db import builders
 from buildbot.db import buildrequests
 from buildbot.db import builds
 from buildbot.db import buildsets
-from buildbot.db import worker
+from buildbot.db import workers
 from buildbot.db import changes
 from buildbot.db import changesources
 from buildbot.db import enginestrategy
@@ -91,7 +91,7 @@ class DBConnector(WorkerAPICompatMixin, service.ReconfigurableServiceMixin,
         self.buildrequests = buildrequests.BuildRequestsConnectorComponent(self)
         self.state = state.StateConnectorComponent(self)
         self.builds = builds.BuildsConnectorComponent(self)
-        self.workers = worker.WorkersConnectorComponent(self)
+        self.workers = workers.WorkersConnectorComponent(self)
         self._registerOldWorkerAttr("workers", name="buildslaves")
         self.users = users.UsersConnectorComponent(self)
         self.masters = masters.MastersConnectorComponent(self)
